scsi: ufs: Enable irq in err_handler
There's a chance that error handler is invoked with
IRQ's disabled. For instance, in case the resume fails
the state is changed to old state - meaning, the state
of the controller is reverted back to suspend.
This entails disabling the IRQ.

Thus, commands timeout during error-handling as well.
Hence, enable IRQs in error-handler before issuing
commands.

CRs-fixed: 2398584
Change-Id: Ia38987ba6791bb74e44b506fda9dc49fd8a32831
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: Sayali Lokhande <sayalil@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>